requirements.txt
# Requirements to run the container
# --------------------------------
psycopg2==2.9.6
pandas==2.0.0
celery==5.2.2
python-dotenv==0.19.0
boto3==1.18.63


